+++ CHANGELOG.md
@@ -23,6 +23,7 @@ Group changes to describe their impact on the project, as follows:
 - Android 10 compatibility
 - New plugin loader to be compatible with app bundle distribution mode
 - Restart service if foreground service setting is on when app is updated
+- Change bluetooth volume while in call if BT device connected and used
 
 ## Changed
 - Improved performances to reduce startup time